pyproject.toml
[tool.poetry]
name = "image-thread-manager"
version = "0.3.0"
description = ""
authors = ["Iris-Raine <iris.raine.darkest.star@gmail.com>"]
readme = "README.md"
packages = [
  {include = "engine"}
]

[tool.poetry.dependencies]
python = "^3.10"
nextcord = "^2.5.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
flake8 = "^6.0.0"
logging = "^0.4.9.6"
pillow = "^10.2.0"

[tool.poetry.scripts]
image-thread-manager-bot = "engine.run:run_discord_bot"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
